feat: add support for string and Unit request and response types
